Move libethcore/Transaction to libethcore/TransactionBase
because it's about class TransactionBase, not Transaction.
There is also libethereum/Transaction.h
