+++ libethereum/Transaction.h
@@ -22,7 +22,7 @@
 #pragma once
 
 #include <libethcore/Common.h>
-#include <libethcore/Transaction.h>
+#include <libethcore/TransactionBase.h>
 #include <libethcore/ChainOperationParams.h>
 #include <libdevcore/RLP.h>
 #include <libdevcore/SHA3.h>
